refactor(auto-connector): reorganize BluetoothAutoConnector file structure
